pyproject.toml
[tool.poetry]
name = "ajax-task"
version = "0.1.0"
description = ""
authors = ["DmitroPodolsky <podolskijdmitro1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
appium-python-client = "^3.1.0"
pydantic = "1.10.13"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
